pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tunnelid-bio"
version = "0.1.0"
description = "Standalone deterministic biometric module for TunnelID"
requires-python = ">=3.10"
dependencies = [
  "numpy==1.26.4",
  "opencv-python==4.10.0.84",
  "mediapipe==0.10.14",
  "insightface==0.7.3",
  "onnxruntime==1.19.2",
  "scikit-learn==1.5.2",
  "scipy==1.14.1",
  "PyYAML==6.0.2",
  "typer==0.12.5",
  "click==8.1.7"
]

[project.scripts]
tunnelid-bio = "tunnelid_bio.cli:app"

[tool.setuptools.packages.find]
include = ["tunnelid_bio*"]
